+++ extensions/audio-device/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Audio Device Changelog
 
+## [Update] - {PR_MERGE_DATE}
+
+- Added combo commands to simultaneously change input and output device
+
 ## [Update] - 2025-02-25
 
 - Added icons for AirPods, AirPlay, and Bluetooth devices